requirements.txt
aiogram
aiosqlite
apscheduler
python-dotenv
aiohttp
